Add support for null in an expression
Initially, a null value/node results in an empty string. In some
occasions this is appropriate (UnaryExpressions - LeftOperand). But in
some occassions it's not (comparing for null).

Hence I made a special object that should be used in the occassion where
a value/node should generate an empty string. Else it results in a
parameter being created. When the query is being prepared, then
parameter that's null should then be replaced with the appropriate null
representation according to the DB of target.